+++ serverside/makoto/src/test/kotlin/KeyConversionTests.kt
@@ -3,7 +3,7 @@ import at.asitplus.attestation.parseToPublicKey
 import at.asitplus.signum.indispensable.toCryptoPublicKey
 import at.asitplus.testballoon.invoke
 import at.asitplus.testballoon.minus
-import de.infix.testBalloon.framework.testSuite
+import de.infix.testBalloon.framework.core.testSuite
 import io.kotest.matchers.shouldBe
 
 val KeyConversionTests by testSuite {